Add first DNN example
This is the first example of DNN created using Tensorflow. As written in the notebook, to attempt this I follow the book "Hands-On Machine Learning" by Aurélien Géron. Right now, I'm interested in understanding better DNNs and Tensorflow library. In the future I'll try to make some kind of interesting stuff.