First try
This is (should be) a Super Mario levels generator. Right now I have some
doubts about the accuracy of this model, but the results are not so bad:
at least it is able to understand the length and to put a final 'floor'
letter. I'm still learning how RNNs work, so I'll probably make some
improvements later.